Frontend lint and format
